feat: типизировать SpellModalForm
